requirements.txt
wheel==0.24.0
Django==1.8.5
psycopg2==2.6.1
djangorestframework==3.3.1
